agent. Remove the dump-env command
- Apply the dump-env behavior automatically at the end of each step.
  Remove the dump-env command.
- Read environment from the state directory at the beginning of each
  step. This allows to pick up changes from other actions running on the
  same agent.
- Fix. Remove variables from Redis HASH with unset-var
- Fix dump-env documentation
