+++ core/agent/agent.go
@@ -26,7 +26,6 @@ import (
 	"log"
 	"os"
 	"os/signal"
-	"strings"
 	"syscall"
 	"time"
 
@@ -66,25 +65,6 @@ func (v *flagStringSlice) Set(raw string) error {
 		*v = append(*v, raw)
 	}
 	return nil
-}
-
-func readEnvironmentFile() []string {
-	env := make([]string, 0)
-	content, err := os.ReadFile("./environment")
-	if err != nil {
-		log.Printf(SD_ERR+"Cannot read ./environment file: %s", err)
-		return env
-	}
-	for _, line := range strings.Split(string(content), "\n") {
-		if line == "" {
-			continue
-		}
-		if strings.Contains(line, "=") == false {
-			continue
-		}
-		env = append(env, line)
-	}
-	return env
 }
 
 func setClientNameCallback(ctx context.Context, cn *redis.Conn) error {
